refactor: move copy logic to entity table rather than related entity!
