+++ src/app/child-dev-project/attendance/activity-attendance-section/activity-attendance-section.component.html
@@ -18,6 +18,7 @@
 </div>
 
 <app-entities-table
+  [mainEntity]="entity"
   [entityType]="entityCtr"
   [records]="records"
   [customColumns]="columns"
