+++ src/app/child-dev-project/children/child-school-overview/child-school-overview.component.ts
@@ -14,7 +14,6 @@ import { EntityMapperService } from "../../../core/entity/entity-mapper/entity-m
 import { EntityRegistry } from "../../../core/entity/database-entity.decorator";
 import { ScreenWidthObserver } from "../../../utils/media/screen-size-observer.service";
 import { FilterService } from "../../../core/filter/filter.service";
-import { EntityActionsService } from "app/core/entity/entity-actions/entity-actions.service";
 
 // TODO: once schema-generated indices are available (#262), remove this component and use its generic super class directly
 @DynamicComponent("ChildSchoolOverview")
@@ -50,14 +49,12 @@ export class ChildSchoolOverviewComponent
     entityRegistry: EntityRegistry,
     screenWidthObserver: ScreenWidthObserver,
     filterService: FilterService,
-    entityActionsService: EntityActionsService,
   ) {
     super(
       entityMapper,
       entityRegistry,
       screenWidthObserver,
       filterService,
-      entityActionsService,
       null,
     );
