+++ src/app/core/entity-details/related-entities-with-summary/related-entities-with-summary.component.html
@@ -1,4 +1,5 @@
 <app-entities-table
+  [mainEntity]="entity"
   [entityType]="entityCtr"
   [records]="data"
   [filter]="filter"
