+++ src/app/core/entity-details/related-entities/related-entities.component.html
@@ -1,17 +1,5 @@
-@if (isCustomFormAvailable) {
-  <button
-    mat-flat-button
-    i18n
-    (click)="copyCustomFormLink()"
-    matTooltip="Copy a public form URL that links this record to the form submission."
-    i18n-matTooltip
-  >
-    Copy custom form ({{ isCustomFormAvailable.title }}) link
-    <fa-icon class="color-accent" size="xl" icon="copy"></fa-icon>
-  </button>
-}
-
 <app-entities-table
+  [mainEntity]="entity"
   [entityType]="entityCtr"
   [records]="data"
   [filter]="filter"
