+++ src/app/core/entity-details/related-time-period-entities/related-time-period-entities.component.html
@@ -5,6 +5,7 @@
 </app-pill>
 
 <app-entities-table
+  [mainEntity]="entity"
   [entityType]="entityCtr"
   [records]="data"
   [filter]="filter"
