+++ src/app/core/entity/entity-actions/entity-actions.service.ts
@@ -11,9 +11,6 @@ import { OkButton } from "../../common-components/confirmation-dialog/confirmati
 import { CascadingActionResult } from "./cascading-entity-action";
 import { EntityActionsMenuService } from "../../entity-details/entity-actions-menu/entity-actions-menu.service";
 import { DuplicateRecordService } from "app/core/entity-list/duplicate-records/duplicate-records.service";
-import { PublicFormConfig } from "app/features/public-form/public-form-config";
-import { AlertService } from "app/core/alerts/alert.service";
-
 /**
  * A service that can triggers a user flow for entity actions (e.g. to safely remove or anonymize an entity),
  * including a confirmation dialog.
@@ -31,7 +28,6 @@ export class EntityActionsService {
     private entityAnonymize: EntityAnonymizeService,
     entityActionsMenuService: EntityActionsMenuService,
     private duplicateRecordService: DuplicateRecordService,
-    private alertService: AlertService,
   ) {
     entityActionsMenuService.registerActions([
       {
@@ -69,7 +65,6 @@ export class EntityActionsService {
         tooltip: $localize`:entity context menu tooltip:Create a copy of this record.`,
       },
     ]);
-    this.initCustomFormActions(entityActionsMenuService);
   }
 
   showSnackbarConfirmationWithUndo(
@@ -314,72 +309,5 @@ export class EntityActionsService {
         entities[0].getConstructor().label
       } "${entities.toString()}" ${action}`;
     }
-  }
-
-  /**
-   * Initializes and registers custom form actions for entities.
-   * - Loads all PublicFormConfig entries from the EntityMapper.
-   * - Filters configs to those with a linkedEntity ID.
-   * - For each matching config, registers a "copy-form-<route>" action that:
-   *   • Executes copying a prebuilt form URL based on the config.
-   *   • Is only visible when matching configs exist for the given entity.
-   */
-  private async initCustomFormActions(
-    entityActionsMenuService: EntityActionsMenuService,
-  ) {
-    const allForms = await this.entityMapper.loadType(PublicFormConfig);
-    const matchingForms = allForms.filter((config) => config.linkedEntity?.id);
-    for (const config of matchingForms) {
-      entityActionsMenuService.registerActions([
-        {
-          action: `copy-form-${config.route}`,
-          execute: (entity) =>
-            this.copyPublicFormLinkFromConfig(entity, config),
-          permission: "read",
-          icon: "link",
-          label: `Copy Custom Form (${config.title})`,
-          tooltip: `Copy a public form URL for ${config.title} that links this entity to the submission.`,
-          visible: (entity) =>
-            this.getMatchingPublicFormConfigs(config, entity),
-        },
-      ]);
-    }
-  }
-
-  /**
-   * Copies the public form link to clipboard if a matching form exists for the given entity.
-   * It checks all PublicFormConfig entries to find the one linked to the current entity type via `linkedEntity.id`.
-   * If a matching form is found, it generates the link including the entity ID as a query parameter and copies it.
-   */
-
-  public async copyPublicFormLinkFromConfig(
-    entity: Entity,
-    config: PublicFormConfig,
-  ): Promise<boolean> {
-    const paramKey = config.linkedEntity.id;
-    const entityId = entity.getId();
-    const fullUrl = `${window.location.origin}/public-form/form/${config.route}?${paramKey}=${encodeURIComponent(entityId)}`;
-
-    await navigator.clipboard.writeText(fullUrl);
-    this.alertService.addInfo("Link copied: " + fullUrl);
-    return true;
-  }
-
-  public getMatchingPublicFormConfigs(
-    config: PublicFormConfig,
-    entity: Entity,
-  ): Promise<boolean> {
-    const entityType = entity.getConstructor().ENTITY_TYPE.toLowerCase();
-    const linkedEntity = config.linkedEntity;
-
-    if (!linkedEntity) return Promise.resolve(false);
-
-    if (linkedEntity.additional) {
-      return Promise.resolve(
-        linkedEntity.additional.toLowerCase() === entityType,
-      );
-    }
-
-    return Promise.resolve(linkedEntity.id.toLowerCase() === entityType);
   }
 }
