+++ src/app/features/matching-entities/matching-entities/matching-entities.component.html
@@ -101,6 +101,7 @@
       ></app-filter>
 
       <app-entities-table
+        [mainEntity]="entity"
         [entityType]="side.entityType"
         [records]="side.availableEntities"
         [customColumns]="side.columns"
