+++ src/app/features/todos/todos-related-to-entity/todos-related-to-entity.component.html
@@ -1,4 +1,5 @@
 <app-entities-table
+  [mainEntity]="entity"
   [entityType]="entityCtr"
   [records]="data"
   [customColumns]="_columns"
